requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black, virtualenv
apscheduler==3.6.3        # via -r requirements.in
astroid==2.4.2            # via pylint
atomicwrites==1.4.0       # via pytest
attrs==20.1.0             # via black, pytest
black==19.10b0            # via -r requirements.in
blinker==1.4              # via raven
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
cffi==1.14.2              # via cryptography
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via black, flask, pip-tools
colorama==0.4.3           # via pylint, pytest
cryptography==3.0         # via -r requirements.in
distlib==0.3.1            # via virtualenv
filelock==3.0.12          # via virtualenv
flask==1.1.2              # via -r requirements.in, raven
google-api-core==1.22.1   # via google-api-python-client
google-api-python-client==1.10.0  # via -r requirements.in
google-auth-httplib2==0.0.4  # via -r requirements.in, google-api-python-client
google-auth-oauthlib==0.4.1  # via -r requirements.in
google-auth==1.20.1       # via google-api-core, google-api-python-client, google-auth-httplib2, google-auth-oauthlib
googleapis-common-protos==1.52.0  # via google-api-core
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
identify==1.4.30          # via pre-commit
idna==2.8                 # via requests
imapclient==2.1.0         # via -r requirements.in
iniconfig==1.0.1          # via pytest
isort==5.4.2              # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
more-itertools==8.4.0     # via pytest
nodeenv==1.5.0            # via pre-commit
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
peewee==3.13.3            # via -r requirements.in
personalcapital==1.0.1    # via -r requirements.in
pip-tools==5.3.1          # via -r requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements.in
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
py==1.9.0                 # via pytest
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pyfakefs==4.1.0           # via -r requirements.in
pylint==2.6.0             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest==6.0.1             # via -r requirements.in
pytz==2020.1              # via apscheduler, google-api-core, tzlocal
pyyaml==5.3.1             # via pre-commit
raven[flask]==6.10.0      # via -r requirements.in
regex==2020.7.14          # via black
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.24.0          # via -r requirements.in, google-api-core, requests-oauthlib
rsa==4.6                  # via google-auth
six==1.15.0               # via apscheduler, astroid, cryptography, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, imapclient, packaging, pip-tools, protobuf, virtualenv
sqlitedict==1.6.0         # via -r requirements.in
toml==0.10.1              # via black, pre-commit, pylint, pytest
typed-ast==1.4.1          # via black
tzlocal==2.1              # via -r requirements.in, apscheduler
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.24.3           # via requests
virtualenv==20.0.31       # via pre-commit
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
